♻️ [Refactoring]: #455 Phase 3 nest カテゴリを tests/nest.rs に分離
バランスネストのテスト 3 件を literal_string/tests/nest.rs に
切り出し、tests.rs ハブに `mod nest;` 宣言を追加。
